🗺️ Atlas: extract context and registry from interpreter blob
