+++ crates/duke-interpreter/src/registry.rs
@@ -1,0 +1,341 @@
+use std::collections::{HashMap, HashSet};
+use std::io::Write;
+
+use duke_loader::ClassLoader;
+use duke_runtime::{Slot, VmError, VmResult};
+
+use crate::context::ClassContext;
+
+/// Metadata for a lambda proxy object created by `LambdaMetafactory`.
+#[derive(Debug, Clone)]
+pub(crate) struct LambdaInfo {
+    pub impl_class: String,
+    pub impl_method: String,
+    pub impl_desc: String,
+    pub impl_kind: u8,
+    pub sam_method: String,
+    #[allow(dead_code)]
+    pub sam_desc: String,
+    pub captured_count: usize,
+}
+
+pub struct ClassRegistry {
+    classes: HashMap<String, ClassContext>,
+    natives: NativeRegistry,
+    /// Tracks which classes have had their `<clinit>` run.
+    initialized: HashSet<String>,
+    /// Lambda proxy class name → metadata.
+    lambdas: HashMap<String, LambdaInfo>,
+    /// Monotonic counter for generating unique lambda class names.
+    lambda_counter: u64,
+    #[cfg(feature = "telemetry")]
+    pub telemetry: duke_telemetry::TelemetryStore,
+}
+
+impl ClassRegistry {
+    /// Creates a new empty class registry.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use duke_interpreter::ClassRegistry;
+    ///
+    /// let mut registry = ClassRegistry::new();
+    /// assert!(!registry.contains("MyClass"));
+    /// ```
+    #[must_use]
+    pub fn new() -> Self {
+        Self {
+            classes: HashMap::new(),
+            natives: NativeRegistry::new(),
+            initialized: HashSet::new(),
+            lambdas: HashMap::new(),
+            lambda_counter: 0,
+            #[cfg(feature = "telemetry")]
+            telemetry: duke_telemetry::TelemetryStore::default(),
+        }
+    }
+
+    pub(crate) fn register_lambda(&mut self, info: LambdaInfo) -> String {
+        let name = format!("$$Lambda${}", self.lambda_counter);
+        self.lambda_counter += 1;
+        self.lambdas.insert(name.clone(), info);
+        name
+    }
+
+    pub(crate) fn get_lambda(&self, class_name: &str) -> Option<&LambdaInfo> {
+        self.lambdas.get(class_name)
+    }
+
+    /// Check if a class has been initialized (clinit has run).
+    #[must_use]
+    pub fn is_initialized(&self, name: &str) -> bool {
+        self.initialized.contains(name)
+    }
+
+    /// Mark a class as initialized.
+    pub fn mark_initialized(&mut self, name: &str) {
+        self.initialized.insert(name.to_string());
+    }
+
+    /// Access the native method registry.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use duke_interpreter::ClassRegistry;
+    ///
+    /// let registry = ClassRegistry::new();
+    /// assert!(registry.natives().get("java/lang/System", "exit", "(I)V").is_none());
+    /// ```
+    #[must_use]
+    pub const fn natives(&self) -> &NativeRegistry {
+        &self.natives
+    }
+
+    /// Access the native method registry mutably.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use duke_interpreter::ClassRegistry;
+    ///
+    /// let mut registry = ClassRegistry::new();
+    /// // registry.natives_mut().register(...)
+    /// ```
+    pub const fn natives_mut(&mut self) -> &mut NativeRegistry {
+        &mut self.natives
+    }
+
+    /// Register a pre-built `ClassContext`.
+    pub fn register(&mut self, ctx: ClassContext) {
+        self.classes.insert(ctx.class_name.clone(), ctx);
+    }
+
+    /// Get a reference to a loaded class.
+    ///
+    /// # Errors
+    /// Returns [`VmError::ClassNotFound`] if the class is not loaded.
+    pub fn get(&self, name: &str) -> VmResult<&ClassContext> {
+        self.classes
+            .get(name)
+            .ok_or_else(|| VmError::ClassNotFound {
+                name: name.to_string(),
+            })
+    }
+
+    /// Get a mutable reference to a loaded class.
+    ///
+    /// # Errors
+    /// Returns [`VmError::ClassNotFound`] if the class is not loaded.
+    pub fn get_mut(&mut self, name: &str) -> VmResult<&mut ClassContext> {
+        self.classes
+            .get_mut(name)
+            .ok_or_else(|| VmError::ClassNotFound {
+                name: name.to_string(),
+            })
+    }
+
+    /// Ensure a class is loaded. If not already present, loads it via the class
+    /// loader, parses it, builds a `ClassContext`, and registers it.
+    ///
+    /// Also recursively loads the superclass chain so that field slot offsets
+    /// can be computed correctly before any object of this type is allocated.
+    ///
+    /// Returns `Ok(true)` if loaded, `Ok(false)` if the class could not be found
+    /// (soft failure — for classes like `java/lang/Object` that we can't load yet).
+    ///
+    /// # Errors
+    /// Returns [`VmError`] if loading the superclass chain fails unexpectedly.
+    pub fn ensure_loaded(&mut self, name: &str, loader: &dyn ClassLoader) -> VmResult<bool> {
+        if self.classes.contains_key(name) {
+            return Ok(true);
+        }
+        let Ok(bytes) = loader.find_class(name) else {
+            return Ok(false);
+        };
+        let Ok(cf) = duke_classfile::parse(&bytes) else {
+            return Ok(false);
+        };
+        let ctx = crate::build_class_context(&cf);
+        let super_class = ctx.super_class.clone();
+        self.classes.insert(name.to_string(), ctx);
+        // Recursively load the superclass so ancestor field counts are known.
+        if let Some(sc) = super_class {
+            self.ensure_loaded(&sc, loader)?;
+        }
+        Ok(true)
+    }
+
+    /// Check if a class is loaded.
+    #[must_use]
+    pub fn contains(&self, name: &str) -> bool {
+        self.classes.contains_key(name)
+    }
+
+    /// Iterate all registered class contexts — used by GC root gathering.
+    pub fn all_classes(&self) -> impl Iterator<Item = &ClassContext> {
+        self.classes.values()
+    }
+
+    /// Mutably iterate all registered class contexts — used to patch static
+    /// field slots after a minor GC collection.
+    pub fn all_classes_mut(&mut self) -> impl Iterator<Item = &mut ClassContext> {
+        self.classes.values_mut()
+    }
+}
+
+impl Default for ClassRegistry {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+/// Signature for native method implementations.
+///
+/// Arguments:
+/// - `&[Slot]`: method arguments (including `this` in slot 0 for instance methods)
+/// - `&mut Heap`: the object heap for reading/writing objects
+/// - `&mut dyn Write`: output sink (stdout in production, `Vec<u8>` in tests)
+pub type NativeHandler = fn(&[Slot], &mut duke_gc::Heap, &mut dyn Write) -> VmResult<Option<Slot>>;
+
+/// A native handler that can call back into the interpreter to invoke Java methods.
+///
+/// The `invoke` closure takes `heap` and `output` as *parameters* (not captured),
+/// using the "loan" pattern: the handler passes its borrows through each call and
+/// gets them back when the call returns. Sequential reborrows — no unsafe required.
+pub type CallbackNativeHandler = fn(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    output: &mut dyn Write,
+    invoke: &mut dyn FnMut(
+        &mut duke_gc::Heap,
+        &mut dyn Write,
+        &str, // class name
+        &str, // method name
+        &str, // descriptor
+        Vec<Slot>,
+    ) -> VmResult<Option<Slot>>,
+) -> VmResult<Option<Slot>>;
+
+/// The `invoke` closure type passed into [`CallbackNativeHandler`] implementations.
+///
+/// Defined separately so function signatures that accept this parameter avoid the
+/// `clippy::type_complexity` lint.
+///
+/// `pub` so that external crates can write their own [`CallbackNativeHandler`]
+/// implementations.  If external registration is not needed, consider
+/// narrowing to `pub(crate)`.
+pub type InvokeFn<'a> = dyn FnMut(&mut duke_gc::Heap, &mut dyn Write, &str, &str, &str, Vec<Slot>) -> VmResult<Option<Slot>>
+    + 'a;
+
+/// Stored in `NativeRegistry` — all existing handlers stay `Simple`.
+#[derive(Copy, Clone, Debug)]
+pub enum HandlerKind {
+    Simple(NativeHandler),
+    Callback(CallbackNativeHandler),
+}
+
+/// Registry of native method implementations.
+///
+/// Maps `"class_name\x00method_name\x00descriptor"` to a handler kind.
+///
+/// # Examples
+///
+/// ```
+/// use duke_interpreter::NativeRegistry;
+///
+/// let mut natives = NativeRegistry::new();
+/// assert!(natives.get("java/lang/System", "exit", "(I)V").is_none());
+/// ```
+pub struct NativeRegistry {
+    handlers: HashMap<String, HandlerKind>,
+}
+
+/// Build the lookup key for a native method: `"class\x00method\x00desc"`.
+///
+/// Using NUL as separator avoids ambiguity (JVM identifiers cannot contain NUL)
+/// and reduces the three allocations previously required per lookup to one.
+#[inline]
+fn make_key(class: &str, method: &str, desc: &str) -> String {
+    let mut key = String::with_capacity(class.len() + method.len() + desc.len() + 2);
+    key.push_str(class);
+    key.push('\x00');
+    key.push_str(method);
+    key.push('\x00');
+    key.push_str(desc);
+    key
+}
+
+impl NativeRegistry {
+    /// Creates a new empty native registry.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use duke_interpreter::NativeRegistry;
+    ///
+    /// let natives = NativeRegistry::new();
+    /// ```
+    #[must_use]
+    pub fn new() -> Self {
+        Self {
+            handlers: HashMap::new(),
+        }
+    }
+
+    fn insert_handler(&mut self, class: &str, method: &str, descriptor: &str, kind: HandlerKind) {
+        self.handlers
+            .insert(make_key(class, method, descriptor), kind);
+    }
+
+    /// Register a native method handler.
+    pub fn register(
+        &mut self,
+        class: &str,
+        method: &str,
+        descriptor: &str,
+        handler: NativeHandler,
+    ) {
+        self.insert_handler(class, method, descriptor, HandlerKind::Simple(handler));
+    }
+
+    /// Register a native method handler that can call back into the interpreter.
+    pub fn register_callback(
+        &mut self,
+        class: &str,
+        method: &str,
+        descriptor: &str,
+        handler: CallbackNativeHandler,
+    ) {
+        self.insert_handler(class, method, descriptor, HandlerKind::Callback(handler));
+    }
+
+    /// Look up a native handler for the given class/method/descriptor.
+    ///
+    /// Returns `Some` only for `Simple` handlers. Use [`get_kind`] to handle
+    /// `Callback` variants.
+    ///
+    /// [`get_kind`]: NativeRegistry::get_kind
+    #[must_use]
+    pub fn get(&self, class: &str, method: &str, descriptor: &str) -> Option<NativeHandler> {
+        match self.handlers.get(&make_key(class, method, descriptor))? {
+            HandlerKind::Simple(h) => Some(*h),
+            HandlerKind::Callback(_) => None,
+        }
+    }
+
+    /// Look up any handler kind for the given class/method/descriptor.
+    #[must_use]
+    pub fn get_kind(&self, class: &str, method: &str, descriptor: &str) -> Option<HandlerKind> {
+        self.handlers
+            .get(&make_key(class, method, descriptor))
+            .copied()
+    }
+}
+
+impl Default for NativeRegistry {
+    fn default() -> Self {
+        Self::new()
+    }
+}
